build passed

+++ techpack/video/msm/vidc/hfi_iris2.c
@@ -154,20 +154,20 @@ void __interrupt_init_iris2(struct venus_hfi_device *device, u32 sid)
 void __setup_ucregion_memory_map_iris2(struct venus_hfi_device *device, u32 sid)
 {
 	__write_register(device, UC_REGION_ADDR_IRIS2,
-			(u32)device->iface_q_table.align_device_addr, sid);
+			(u32)(uintptr_t)device->iface_q_table.align_device_addr, sid);
 	__write_register(device, UC_REGION_SIZE_IRIS2, SHARED_QSIZE, sid);
 	__write_register(device, QTBL_ADDR_IRIS2,
-			(u32)device->iface_q_table.align_device_addr, sid);
+			(u32)(uintptr_t)device->iface_q_table.align_device_addr, sid);
 	__write_register(device, QTBL_INFO_IRIS2, 0x01, sid);
 	if (device->sfr.align_device_addr)
 		__write_register(device, SFR_ADDR_IRIS2,
-				(u32)device->sfr.align_device_addr, sid);
+				(u32)(uintptr_t)device->sfr.align_device_addr, sid);
 	if (device->qdss.align_device_addr)
 		__write_register(device, MMAP_ADDR_IRIS2,
-				(u32)device->qdss.align_device_addr, sid);
+				(u32)(uintptr_t)device->qdss.align_device_addr, sid);
 	/* update queues vaddr for debug purpose */
 	__write_register(device, CPU_CS_VCICMDARG0_IRIS2,
-		(u32)device->iface_q_table.align_virtual_addr, sid);
+		(u32)(uintptr_t)device->iface_q_table.align_virtual_addr, sid);
 	__write_register(device, CPU_CS_VCICMDARG1_IRIS2,
 		(u32)((u64)device->iface_q_table.align_virtual_addr >> 32),
 		sid);
